+++ lib/src/views/flutter_native_ad_view.dart
@@ -106,6 +106,8 @@ class _FlutterNativeAdViewState extends State<FlutterNativeAdView> {
         placementId: widget.placementId,
         adType: widget.adType,
         position: widget.position,
+        nativeLogicalWidth: widget.width.round(),
+        nativeLogicalHeight: widget.height.round(),
       );
 
       if (apiURL == null) {